feat: stripe integration (#17)
* feat(payment): ❇️ add StripePaymentProvider for handling payments via Stripe API

- Introduced a new StripePaymentProvider to facilitate payment processing using the Stripe API.
- Implemented methods for initiating payments and retrieving payment statuses, including error handling and logging.
- Updated go.mod and go.sum to include the Stripe Go library version 82.3.0 for integration.
- Added necessary imports and structured the provider to align with existing payment processing architecture.

* refactor(payment): 🛠️ update StripePaymentProvider to use Stripe client for payment operations

- Refactored the StripePaymentProvider to utilize a Stripe client instead of directly setting the API key.
- Updated methods for initiating payments and retrieving payment statuses to align with the new client structure.
- Enhanced error handling and logging for payment operations, improving maintainability and clarity.

* feat(payment): ✨ integrate Stripe payment provider and enhance payment handling

- Added a new Stripe payment provider to facilitate payment processing via the Stripe API.
- Updated the application configuration to include payment provider settings in the .env_sample file.
- Refactored account handling to incorporate payment provider dependencies, improving the deposit and withdrawal processes.
- Enhanced error handling and logging for payment operations, ensuring better traceability and maintainability.
- Updated tests to validate the integration of the new payment provider and its functionalities.

* docs(stripe): 📚 add comprehensive documentation for Stripe integration and multi-currency deposit refactor

- Introduced a new documentation file detailing the integration of Stripe as a payment provider and the refactoring of the multi-currency deposit flow.
- Documented improvements in exchange rate caching logic and handling of zero-decimal currencies, addressing critical bugs and enhancing auditability.
- Outlined the problems faced during implementation and the solutions applied, ensuring clarity for future developers.
- Included next steps for further integration and testing, along with references to relevant SDKs and project files.

* refactor(config): 🔧 migrate configuration imports to a unified package

- Replaced all instances of the old config package import path with the new unified config package.
- Updated various files across the application to ensure consistent usage of the new configuration structure.
- Removed the deprecated pkg/config package and its associated tests to streamline the codebase.
- Enhanced clarity and maintainability by consolidating configuration management into a single package.

* feat(docs): 📚 add MkDocs configuration and comprehensive documentation for fintech platform

- Introduced a new MkDocs configuration file to set up the documentation site for the fintech platform.
- Created detailed documentation covering the architecture, features, and usage of the application, including sections on currency management, payments, and refactoring patterns.
- Added navigation structure to enhance user experience and accessibility of documentation.
- Included OpenAPI specifications for the API reference, ensuring clarity for developers and contributors.
- Established a blog section for ongoing updates and insights related to the fintech platform.
- Revised  to provide a concise overview of the project's architecture, emphasizing the event-driven design and separation of concerns.
- Introduced a new  file to guide contributors on best practices, including branching, commit message formatting, and testing requirements.
- Enhanced clarity and organization of documentation to support better onboarding and understanding for new contributors.
- Updated various sections to reflect recent architectural changes and improvements in the codebase.

* feat(docs): 📚 enhance documentation structure and content for better onboarding

- Updated the MkDocs configuration to include a new "Getting Started" section in the navigation, improving accessibility for new users.
- Added an icon to the "Getting Started" page and revised its content to provide clearer guidance on setting up the Fintech Platform.
- Refined the main documentation index to better reflect the project's purpose and technologies used, enhancing clarity and organization.
- Introduced a table summarizing the technologies utilized in the project, providing quick reference for developers.
- Improved overall documentation layout and added visual elements to facilitate better understanding of the platform's architecture and features.

* ci(docs): 👷 add GitHub Actions workflow for documentation deployment

- Introduced a new GitHub Actions workflow to automate the deployment of documentation to GitHub Pages.
- Configured steps for checking out the repository, setting up Python, caching dependencies, and deploying with MkDocs.
- This enhancement streamlines the documentation update process, ensuring that changes are reflected promptly on the documentation site.

+++ infra/provider/mock_payment_provider.go
@@ -38,7 +38,7 @@ func NewMockPaymentProvider() *MockPaymentProvider {
 }
 
 // InitiatePayment simulates initiating a deposit payment.
-func (m *MockPaymentProvider) InitiatePayment(ctx context.Context, userID, accountID uuid.UUID, amount float64, currency string) (string, error) {
+func (m *MockPaymentProvider) InitiatePayment(ctx context.Context, userID, accountID uuid.UUID, amount int64, currency string) (string, error) {
 	paymentID := uuid.New().String()
 	m.mu.Lock()
 	m.payments[paymentID] = &mockPayment{status: provider.PaymentPending}
