+++ pkg/provider/payment.go
@@ -20,6 +20,6 @@ const (
 
 // PaymentProvider is a interface for payment provider
 type PaymentProvider interface {
-	InitiatePayment(ctx context.Context, userID, accountID uuid.UUID, amount float64, currency string) (string, error)
+	InitiatePayment(ctx context.Context, userID, accountID uuid.UUID, amount int64, currency string) (string, error)
 	GetPaymentStatus(ctx context.Context, paymentID string) (PaymentStatus, error)
 }
